fix(menu): include manquants

+++ programme/src/menu.c
@@ -2,7 +2,7 @@
 
 int obtenirProfondeurIA(int nbArguments, char ** arguments) {
   if (nbArguments == 3){
-    return arguments[2];
+    return (int)arguments[2];
   }
   return PROFONDEUR_DEFAUT_IA;
 }
